backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
graphene-django==3.1.5
django-filter==23.3
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-environ==0.11.2
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
django-graphql-jwt==0.4.0
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
black==23.11.0
flake8==6.1.0
isort==5.12.0

